[SYCL] Update test for ext_sycl_oneapi_clock
Add a test plan and update the test
